+++ src/qianfan/resources/console/utils.py
@@ -16,10 +16,10 @@
 Utils for console api
 """
 import functools
-from typing import Any, Callable
+from typing import Any, Awaitable, Callable
 
 from qianfan.consts import DefaultValue
-from qianfan.resources.api_requestor import ConsoleAPIRequestor
+from qianfan.resources.requestor.console_requestor import ConsoleAPIRequestor
 from qianfan.resources.typing import ParamSpec, QfRequest, QfResponse, RetryConfig
 from qianfan.utils import _get_console_ak_sk
 
@@ -51,6 +51,38 @@ def console_api_request(func: Callable[P, QfRequest]) -> Callable[P, QfResponse]
             ),
         )
         req = func(*args, **kwargs)
+        return ConsoleAPIRequestor()._request_console_api(req, ak, sk, retry_config)
+
+    return inner
+
+
+def async_console_api_request(
+    func: Callable[P, Awaitable[QfRequest]]
+) -> Callable[P, Awaitable[QfResponse]]:
+    """
+    wrapper for all functions in sdk for console api, so that the function
+    only needs to provide the request
+    this decorator will:
+    1. extract ak and sk from kwargs
+    2. extract retry config from kwargs
+    3. use the requestor to send request
+    4. return the response to the user
+    """
+
+    @functools.wraps(func)
+    async def inner(*args: Any, **kwargs: Any) -> QfResponse:
+        """
+        inner function of wrapper
+        """
+        ak, sk = _get_console_ak_sk(**kwargs)
+        retry_config = RetryConfig(
+            retry_count=kwargs.get("retry_count", DefaultValue.RetryCount),
+            timeout=kwargs.get("request_timeout", DefaultValue.RetryTimeout),
+            backoff_factor=kwargs.get(
+                "backoff_factor", DefaultValue.RetryBackoffFactor
+            ),
+        )
+        req = await func(*args, **kwargs)
         return ConsoleAPIRequestor()._request_console_api(req, ak, sk, retry_config)
 
     return inner
